+++ lib/speech_to_text_v2.dart
@@ -1,0 +1,135 @@
+library flutter_google_speech;
+
+import 'dart:async';
+
+import 'package:google_speech/auth/third_party_authenticator.dart';
+import 'package:google_speech/generated/google/cloud/speech/v2/cloud_speech.pb.dart'
+    hide RecognitionConfig, StreamingRecognitionConfig;
+import 'package:google_speech/generated/google/cloud/speech/v2/cloud_speech.pbgrpc.dart'
+    hide RecognitionConfig, StreamingRecognitionConfig;
+import 'package:google_speech/speech_client_authenticator.dart';
+import 'package:grpc/grpc.dart';
+
+import 'config/recognition_config_v2.dart';
+import 'config/streaming_recognition_config.dart';
+import 'generated/google/longrunning/operations.pb.dart';
+
+/// An interface to Google's Speech-to-Text Api via grpc.
+///
+/// Important! You need an own Google project to use this class.
+class SpeechToTextV2 {
+  /// [CallOptions] that are transferred when creating a [SpeechClient].
+  final CallOptions _options;
+
+  /// Your Google Project Id. This is required for Version2
+  final String projectId;
+
+  /// [ClientChannel] which is used for the Google Speech-to-Text Api.
+  final ClientChannel _channel = ClientChannel('speech.googleapis.com');
+
+  // Private constructor to prevent direct initialization of the class.
+  SpeechToTextV2._(this._options, {required this.projectId});
+
+  /// Creates a SpeechToTextV2 interface using a service account.
+  factory SpeechToTextV2.viaServiceAccount(ServiceAccount account) =>
+      SpeechToTextV2._(account.callOptions, projectId: account.projectId!);
+
+  /// Creates a SpeechToTextV2 interface using a third party authenticator.
+  /// Don't worry about updating the access token, the package does it automatically.
+  /// Example:
+  ///       final SpeechToTextV2 = SpeechToTextV2.viaThirdPartyAuthenticator(
+  ///         ThirdPartyAuthenticator(
+  ///           obtainCredentialsFromThirdParty: () async {
+  ///             // request api to get token
+  ///             final json = await requestCredentialFromMyApi();
+  ///             return AccessCredentials.fromJson(json);
+  ///           },
+  ///         ),
+  ///       );
+  factory SpeechToTextV2.viaThirdPartyAuthenticator(
+          ThirdPartyAuthenticator thirdPartyAuthenticator,
+          {required String projectId}) =>
+      SpeechToTextV2._(thirdPartyAuthenticator.toCallOptions,
+          projectId: projectId);
+
+  /// Creates a SpeechToTextV2 interface using a token.
+  /// You are responsible for updating the token when it expires.
+  factory SpeechToTextV2.viaToken(String typeToken, String token,
+          {required String projectId}) =>
+      SpeechToTextV2._(
+          CallOptions(metadata: {'authorization': '$typeToken $token'}),
+          projectId: projectId);
+
+  /// Listen to audio stream.
+  /// Cancelled as soon as dispose is called.
+  StreamSubscription<List<int>>? _audioStreamSubscription;
+
+  /// Sends a [RecognizeRequest] request to the Google Speech Api.
+  /// Requires a [RecognitionConfig] and an [RecognitionAudio].
+  ///
+  /// Audio files transcribed with recognize must not be longer than 60 seconds.
+  /// For longer audio files [longRunningRecognize] must be used.
+  Future<RecognizeResponse> recognize(
+      RecognitionConfigV2 config, List<int> audio) {
+    final client = SpeechClient(_channel, options: _options);
+
+    // Create the request, which transmits the necessary
+    // data to the Google Api.
+    final request = (RecognizeRequest()
+      ..recognizer = 'projects/$projectId/locations/global/recognizers/_'
+      ..config = config.toConfig()
+      ..content = audio);
+    return client.recognize(request);
+  }
+
+  /// Sends a [StreamingRecognizeRequest] to the Google Speech Api.
+  /// Requires a [StreamingRecognitionConfig] and an audioStream.
+  Stream<StreamingRecognizeResponse> streamingRecognize(
+      StreamingRecognitionConfigV2 config, Stream<List<int>> audioStream) {
+    final client = SpeechClient(_channel, options: _options);
+
+    // Create the stream, which later transmits the necessary
+    // data to the Google Api.
+    final request = StreamController<StreamingRecognizeRequest>();
+
+    // Send the streaming config at first.
+    request.add(StreamingRecognizeRequest()
+      ..streamingConfig = config.toConfig()
+      ..recognizer = 'projects/$projectId/locations/global/recognizers/_');
+
+    _audioStreamSubscription = audioStream.listen((audio) {
+      // Add audio content when stream changes.
+      request.add(StreamingRecognizeRequest()..audio = audio);
+    });
+
+    _audioStreamSubscription!.onDone(() {
+      // Close the request stream, if the audio stream is finished.
+      request.close();
+    });
+    return client.streamingRecognize(request.stream);
+  }
+
+  /// Sends a [LongRunningRecognizeRequest] request to the Google Speech Api.
+  /// Requires a [RecognitionConfigBeta] and an [RecognitionAudio].
+  ///
+  /// To use asynchronous speech recognition to transcribe audio longer than 60
+  /// seconds, you must have your data saved in a Google Cloud Storage bucket.
+  ResponseFuture<Operation> longRunningRecognize(
+      RecognitionConfigV2 config, String audioGcsUri) {
+    final client = SpeechClient(_channel, options: _options);
+
+    // transform audio to RecognitionAudio
+    final recognitionAudio = BatchRecognizeFileMetadata()..uri = audioGcsUri;
+
+    // Create the request, which transmits the necessary
+    // data to the Google Api.
+    final request = BatchRecognizeRequest(
+        files: [recognitionAudio], config: config.toConfig());
+
+    return client.batchRecognize(request);
+  }
+
+  void dispose() {
+    _audioStreamSubscription?.cancel();
+  }
+}